Reduce delta between master and java7
- Reformat code
- rearranged code
- organized imports

+++ openpdf/src/main/java/com/lowagie/text/pdf/PdfDocument.java
@@ -1459,7 +1459,7 @@ public class PdfDocument extends Document {
                     if (chunk.isTab()) {
                         Object[] tab = (Object[])chunk.getAttribute(Chunk.TAB);
                         DrawInterface di = (DrawInterface)tab[0];
-                        tabPosition = tab[1] + tab[3];
+                        tabPosition = (Float) tab[1] + (Float) tab[3];
                         float fontSize = chunk.font().size();
                         float ascender = chunk.font().getFont().getFontDescriptor(BaseFont.ASCENT, fontSize);
                         float descender = chunk.font().getFont().getFontDescriptor(BaseFont.DESCENT, fontSize);
